move all storage related encodings to coding.py
