+++ echopype/utils/coding.py
@@ -74,3 +74,75 @@ def set_encodings(ds: xr.Dataset) -> xr.Dataset:
             new_ds[var].encoding = encoding
 
     return new_ds
+
+
+def get_zarr_compression(var: xr.Variable, compression_settings: dict) -> dict:
+    """Returns the proper zarr compressor for a given variable type"""
+
+    if np.issubdtype(var.dtype, np.floating):
+        return compression_settings["float"]
+    elif np.issubdtype(var.dtype, np.integer):
+        return compression_settings["int"]
+    elif np.issubdtype(var.dtype, np.str_):
+        return compression_settings["string"]
+    elif np.issubdtype(var.dtype, np.datetime64):
+        return compression_settings["time"]
+    else:
+        raise NotImplementedError(f"Zarr Encoding for dtype = {var.dtype} has not been set!")
+
+
+def set_zarr_encodings(ds: xr.Dataset, compression_settings: dict) -> dict:
+    """
+    Obtains all variable encodings based on zarr default values
+    """
+
+    # create zarr specific encoding
+    encoding = dict()
+    for name, val in ds.variables.items():
+
+        val_encoding = val.encoding
+        val_encoding.update(get_zarr_compression(val, compression_settings))
+        encoding[name] = val_encoding
+
+    return encoding
+
+
+def set_netcdf_encodings(ds: xr.Dataset, compression_settings: dict) -> dict:
+    """
+    Obtains all variable encodings based on netcdf default values
+    """
+
+    # TODO: below is the encoding we were using for netcdf, we need to make
+    #  sure that the encoding is appropriate for all data variables
+    encoding = (
+        {var: compression_settings for var in ds.data_vars}
+        if compression_settings is not None
+        else {}
+    )
+
+    return encoding
+
+
+def set_storage_encodings(ds: xr.Dataset, compression_settings: dict, engine: str) -> dict:
+    """
+    Obtains the appropriate zarr or netcdf specific encodings for
+    each variable in ``ds``.
+    """
+
+    if compression_settings is not None:
+
+        if engine == "zarr":
+
+            encoding = set_zarr_encodings(ds, compression_settings)
+
+        elif engine == "netcdf4":
+
+            encoding = set_netcdf_encodings(ds, compression_settings)
+
+        else:
+            raise RuntimeError(f"Obtaining encodings for the engine {engine} is not allowed.")
+
+    else:
+        encoding = dict()
+
+    return encoding
